Rich/fix tests

+++ test/src/Test62.ts
@@ -8,25 +8,20 @@ import { SGAgentUtils } from './utils/SGAgentUtils';
 import { ScriptType, TaskDefTarget } from '../../server/src/shared/Enums';
 import * as _ from 'lodash';
 
-
 const script1 = `
 console.log('Hello from lambda!');
 `;
 const script1_b64 = SGUtils.btoa(script1);
 
-
 let self: Test62;
 
-
 export default class Test62 extends TestBase.WorkflowTestBase {
-
     constructor(testSetup) {
         super('Test62', testSetup);
         this.description = 'Run node task as lambda zipfile loaded to s3';
 
         self = this;
     }
-
 
     public async RunTest() {
         let result: boolean;
@@ -36,17 +31,20 @@ export default class Test62 extends TestBase.WorkflowTestBase {
 
         const workingDirectory: string = process.cwd() + '/test62lambda';
 
-        if (fs.existsSync(workingDirectory))
-            fse.removeSync(workingDirectory);
+        if (fs.existsSync(workingDirectory)) fse.removeSync(workingDirectory);
         fs.mkdirSync(workingDirectory);
 
-        const lambdaZipFile = (<string>await SGAgentUtils.CreateAWSLambdaZipFile_NodeJS(workingDirectory, script1, '', 'test62'));
+        const lambdaZipFile = <string>(
+            await SGAgentUtils.CreateAWSLambdaZipFile_NodeJS(workingDirectory, script1, '', 'test62')
+        );
 
-        let res: any = await self.testSetup.RestAPICall('artifact', 'POST', _teamId, null, {name: path.basename(lambdaZipFile), type: 'multipart/form-data'});
+        let res: any = await self.testSetup.RestAPICall('artifact', 'POST', _teamId, null, {
+            name: path.basename(lambdaZipFile),
+            type: 'multipart/form-data',
+        });
         const artifact = res.data.data;
 
         await SGAgentUtils.RunCommand(`curl -v --upload-file "${lambdaZipFile}" "${artifact.url}"`, {});
-
 
         // const sgAdminTeam: string = config.get('sgAdminTeam');
         // const awsLambdaRequiredTags: string = config.get('awsLambdaRequiredTags');
@@ -59,8 +57,8 @@ export default class Test62 extends TestBase.WorkflowTestBase {
                     name: 'Script 62',
                     scriptType: ScriptType.NODE,
                     code: script1_b64,
-                    shadowCopyCode: script1_b64
-                }
+                    shadowCopyCode: script1_b64,
+                },
             ],
             jobDefs: [
                 {
@@ -77,13 +75,13 @@ export default class Test62 extends TestBase.WorkflowTestBase {
                                     lambdaRole: config.get('lambda-admin-iam-role'),
                                     lambdaAWSRegion: config.get('AWS_REGION'),
                                     lambdaZipfile: artifact.id,
-                                    lambdaFunctionHandler: 'index.handler'
-                                }
-                            ]
-                        }
-                    ]
-                }
-            ]
+                                    lambdaFunctionHandler: 'index.handler',
+                                },
+                            ],
+                        },
+                    ],
+                },
+            ],
         };
 
         const { scripts, jobDefs } = await this.CreateJobDefsFromTemplates(properties);
@@ -91,7 +89,10 @@ export default class Test62 extends TestBase.WorkflowTestBase {
         let startedJobId;
         resApiCall = await this.testSetup.RestAPICall(`job`, 'POST', _teamId, { jobDefName: jobDefs['Job 62'].name });
         if (resApiCall.data.statusCode != 201) {
-            self.logger.LogError('Failed', { Message: `job POST returned ${resApiCall.data.statusCode}`, _jobDefId: jobDefs['Job 62'].id });
+            self.logger.LogError('Failed', {
+                Message: `job POST returned ${resApiCall.data.statusCode}`,
+                _jobDefId: jobDefs['Job 62'].id,
+            });
             return false;
         }
         startedJobId = resApiCall.data.data.id;
@@ -99,40 +100,37 @@ export default class Test62 extends TestBase.WorkflowTestBase {
         const jobStartedBP: any = {
             domainType: 'Job',
             operation: 1,
-            model:
-            {
+            model: {
                 _teamId: config.get('sgTestTeam'),
                 _jobDefId: jobDefs[properties.jobDefs[0].name].id,
                 runId: 0,
                 name: properties.jobDefs[0].name,
                 status: 0,
                 id: startedJobId,
-                type: 'Job'
-            }
-        }
+                type: 'Job',
+            },
+        };
         self.bpMessagesExpected.push(jobStartedBP);
 
         const jobCompletedBP: any = {
             domainType: 'Job',
             operation: 2,
-            model:
-            {
+            model: {
                 status: 20,
                 id: startedJobId,
-                type: 'Job'
-            }
+                type: 'Job',
+            },
         };
         self.bpMessagesExpected.push(jobCompletedBP);
 
         result = await self.WaitForTestToComplete();
-        if (!result)
-            return result;
+        if (!result) return result;
 
         // newAgent.offline = true;
         // await newAgent.SendHeartbeat(false, true);
         // await newAgent.Stop();
 
-        await self.testSetup.RestAPICall(`artifact/${artifact.id}`, 'DELETE', _teamId, null, null)
+        await self.testSetup.RestAPICall(`artifact/${artifact.id}`, 'DELETE', _teamId, null, null);
         fs.unlinkSync(lambdaZipFile);
 
         return true;
